Move patch deployment methods to the wait module
