+++ test/e2e/modules/wait/deployment.go
@@ -11,12 +11,17 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	appsv1 "k8s.io/api/apps/v1"
 	v1 "k8s.io/api/core/v1"
+	"k8s.io/client-go/kubernetes"
 
 	"github.com/NVIDIA/KAI-scheduler/test/e2e/modules/wait/watcher"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/watch"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 	runtimeClient "sigs.k8s.io/controller-runtime/pkg/client"
 )
+
+type ArgsUpdater func(args []string) []string
 
 // WaitForDeploymentPodsRunning waits until all pods from a deployment are in the running state
 func WaitForDeploymentPodsRunning(ctx context.Context, client runtimeClient.WithWatch, name, namespace string) {
@@ -71,4 +76,60 @@ func WaitForDeploymentPodsRunning(ctx context.Context, client runtimeClient.With
 	if !watcher.ForEvent(ctx, client, pw) {
 		Fail(fmt.Sprintf("timed out waiting for pods of deployment %s/%s to be running", namespace, name))
 	}
+}
+
+func PatchDeploymentArgs(
+	ctx context.Context,
+	clientset kubernetes.Interface,
+	namespace string,
+	deploymentName string,
+	containerName string,
+	argsUpdater ArgsUpdater,
+) error {
+	// Get the deployment
+	deployment, err := clientset.AppsV1().Deployments(namespace).Get(
+		ctx,
+		deploymentName,
+		metav1.GetOptions{},
+	)
+	if err != nil {
+		return fmt.Errorf("failed to get deployment %s: %v", deploymentName, err)
+	}
+
+	// Create a copy to use as the base for our patch
+	deploymentCopy := deployment.DeepCopy()
+
+	// Find and update the container args in our copy
+	containerFound := false
+	for i, container := range deploymentCopy.Spec.Template.Spec.Containers {
+		if container.Name == containerName {
+			deploymentCopy.Spec.Template.Spec.Containers[i].Args = argsUpdater(container.Args)
+			containerFound = true
+			break
+		}
+	}
+
+	if !containerFound {
+		return fmt.Errorf("container %s not found in deployment %s", containerName, deploymentName)
+	}
+
+	// Create the patch data
+	patchBytes, err := client.MergeFrom(deployment).Data(deploymentCopy)
+	if err != nil {
+		return fmt.Errorf("failed to create patch data: %v", err)
+	}
+
+	// Apply the patch
+	_, err = clientset.AppsV1().Deployments(namespace).Patch(
+		ctx,
+		deploymentName,
+		types.MergePatchType,
+		patchBytes,
+		metav1.PatchOptions{},
+	)
+	if err != nil {
+		return fmt.Errorf("failed to patch deployment %s: %w", deploymentName, err)
+	}
+
+	return nil
 }
